EOS-26090: Changed url options to update-resource, modified scripts as per review

+++ performance/PerfLine/roles/perfline_setup/files/wrapper/scripts/LC/update.sh
@@ -24,13 +24,15 @@ function deploy_cluster() {
     ssh $PRIMARY_NODE "cd $K8S_SCRIPTS_DIR && ./deploy-cortx-cloud.sh"
 }
 
-function check_docker_image_version() {
+function log_docker_image_version() {
     echo "Docker Image Version"
+    set +e
     pdsh -S -w $NODES 'docker images | grep "cortx-all"'
+    set -e
 }
 
 function install_custom_docker_image() {
-    pdsh -S -w $NODES "docker pull $URL"
+    pdsh -S -w $NODES "docker pull $IMAGE"
 }
 
 function docker_image_cleanups() {
@@ -38,10 +40,15 @@ function docker_image_cleanups() {
     do
         image_id=($(ssh $NODE "docker images | grep cortx-all" | awk '{ print $3 }'))
         echo "image_ids: ${image_id[@]}"
-        for id in ${image_id[@]}
-        do
-           ssh $NODE "docker rmi $id"
-        done
+        if [[ -n "$image_id" ]];
+        then
+            for id in ${image_id[@]}
+            do
+               ssh $NODE "docker rmi $id"
+            done
+        else
+            echo "$NODE : docker images not found"
+        fi
     done
     pdsh -S -w $NODES "rm -rf /mnt/fs-local-volume/*"
   
@@ -55,10 +62,10 @@ fname = "$K8S_SCRIPTS_DIR/solution.yaml"
 
 with open(fname, 'r') as f:
     data = yaml.safe_load(f)
-data['solution']['images']['cortxcontrolprov'] = "$URL"
-data['solution']['images']['cortxcontrol'] = "$URL"
-data['solution']['images']['cortxdataprov'] = "$URL"
-data['solution']['images']['cortxdata'] = "$URL"
+data['solution']['images']['cortxcontrolprov'] = "$IMAGE"
+data['solution']['images']['cortxcontrol'] = "$IMAGE"
+data['solution']['images']['cortxdataprov'] = "$IMAGE"
+data['solution']['images']['cortxdata'] = "$IMAGE"
 
 with open(fname, 'w') as yaml_file:
     yaml_file.write( yaml.dump(data, sort_keys = False))
@@ -73,18 +80,34 @@ echo $output
 function main() {
     echo $@
     destroy_cluster
-    check_docker_image_version
+    log_docker_image_version
 
-    update_solution_file
-    
-    if [[ -n "$URL" ]]; then
+# Below if else statement has been considered for 2 options, 
+# 1st for docker images and 2nd for custom docker images, 
+# else part will be cover by custom docker images option
+
+    if [[ -n "$IMAGE" ]]; then
+        update_solution_file
         docker_image_cleanups
         install_custom_docker_image
+    else
+        echo "This option will be cover by custom docker build"
     fi
 
-    check_docker_image_version
+    log_docker_image_version
     deploy_cluster
 
+}
+
+function validate() {
+
+    if [[ -z "$IMAGE" ]]; then
+       echo "Docker Images are not specified"
+    fi
+    if [[ -z "$NODES" ]]; then
+       echo "Docker Images are not specified"
+    fi
+    
 }
 
 function check_arg_count() {
@@ -103,18 +126,19 @@ while [[ $# -gt 0 ]]; do
             EX_SRV="pdsh -S -w $NODES"
             shift
             ;;
-	--url)
-            URL=$2
+	--update-resource)
+            IMAGE=$2
             shift
             ;;
 
         *)
-            echo -e "Invalid option: $1\nUse --help option"
+            echo -e "Invalid option: $1\n"
             exit 1
             ;;
     esac
     shift
 done
 
+validate
 main
